magit-patch-popup: new popup
Please note that this is an early version.  Future versions will also
feature `git send-email' and `git imap-send' and more arguments.  But
implementing that requires much more work, mainly because it has to
be well integrated with external tools, and won't be done before the
release.

If I think that some important argument is missing, then please add
it locally using `magit-define-popup-{switch,option}'.