magit-git-string: report non-zero exit status
Rename the existing `magit-git-string' to `magit-git-str'; it should
only be used when a non-zero exit status is expected and does not
constitute an error.

Add a new `magit-git-string' which uses `magit-git-insert' and, when
`magit-git-debug' is non-nil, therefor reports a non-zero exit status
in the message area as well as by creating a section with Git's stderr
in the process buffer.

Most callers of `magit-git-string' keep calling that function, i.e
switch to the new implementation; they start reporting errors.  Note
that this does not cause an Elisp error to be raised.  The purpose
of this change is to make it easier to debug, and not to gracefully
recover from errors.

Likewise add `magit-rev-parse-safe' as a variant of `magit-rev-parse'.