magit-get-tracked-{branch,ref}: assume sane configuration
The default value of the Git variable `remote.<remote>.fetch' is
"+refs/heads/*:refs/remotes/<remote>/*".  It can be useful to add
_additional_ values, e.g. to fetch notes.  However it does _not_
make sense to change the default value used for fetching branches.
But it's possible and Git seems to deal with it just fine.  These
functions used to do the same, but it's to expensive - and if a
user actually changes this then she has nobody but herself to
blame.