fix various diff scope related issues
Remove `magit-diff-scope's SINGULAR argument, if that actually behaves
as it was already previously described in the doc-string when SECTION
is non-nil, then this is not necessary.  I.e. when SECTION is specified
return the type of that, not that of a possible selection.

As a result, also treat the `region' type as a singular, because that's
what it is.  `list' is a singular to, by the way.  Only `hunks' and
`files' are not, the presence or absence of the "s" suffix helps
remembering what is plural and what not.

Make `magit-section-internal-region-p's SECTION argument optional, and
when it is present actually use it.  Also make sure the region is inside
SECTION (or the current section if that is nil).  Previously we instead
did something strange in `magit-diff-scope', which resulted in `region'
being treated as a plural with the respective singular being `hunk'
which isn't correct.

The bugs being fixed by all this is that previously `magit-diff-scope'
returned `files' when it should have returned `files' (this was the case
in `magit-discard-apply' and `magit-reverse-apply') and that this bug is
fixed in a way that does not introduce a new bug where `magit-diff-scope'
returned `hunk' instead of `region', as the previous attempt at fixing
the first bug using c3535abe38e2cee9b65d7ef798a22989d1eb0419 did.