magit-insert-tracked-files: new section inserter
By default this is not added to `magit-status-sections-hook' because
one usually doesn't act all that often tracked but unmodified files.
So it is better in my eyes, if it's slightly inconvenient files, than
it is to always waste time refreshing that list when doing anything
that causes a refresh of the status buffer.