magit-read-range-or-commit: provide completion for second rev
This uses `crm' whose purpose actually is to complete an arbitrary
number of items, not one or two as we need here.  So users could insert
something like "a..b..c", but they are not gonna do that, I hope.  Also
after inserting e.g. "master" this claims that's the only completion; it
would be better if it offered "master", "master..", and "master...".
For now this only offers completions for the second revision once ".."
or "..." has been inserted explicitly.

`pcomplete' would be more suitable, but as far as I understand it this
is only for in-buffer completion, not minibuffer completion.