put popup options in correct groups, suppress unnecessary options
Disband the `magit-popups' Custom group and instead place most popup
argument options into `magit-commands'.  The options of popups which
are used to create or update buffers whose major-mode is a Magit mode
are instead placed in the respective mode group.

Also suppress the creation of popup argument options for popups which
do not take any arguments.