do not revert buffers in which magit-blame-mode is enabled
When a buffer, in which `magit-blame-mode' is enabled, was reverted due
to `magit-auto-revert-mode' being on, this was done by first disabling
`magit-blame-mode, then reverting, and finally enabling the mode again.

Until 0576e69b3f2299a3a47332116b4c6250d3fe6be4 the same was done for
`auto-revert-mode' too.

Unfortunately turning the mode on again, caused the file-visiting buffer
to be also displayed in the window that was selected when the action
which triggered the refresh (and thus the revert).  The reason for this
is unknown, see #1731.

To work around this keep `magit-auto-revert-mode' from reverting when
`magit-blame-mode' is on.  Keep doing so for `auto-revert-mode'.  In
both cases how show a warning.