Add magit-send-email
Per #1747

Completes for a rev range but needs to also accept files and directories
as well for full support.

Git `send-email` requires the `--to` field, and will prompt for it if
unspecified, so it gets stuck, but there should be a way to fix this. We
need to prompt for the "to" option if not given.

Git requires specifying multiple address with repeated uses of the
option, like ``--from=a@b.c" --from=x@y.z``. I'm not sure how to do that
from within magit-popup. Again we could prompt for each, but I'd really
rather not.

Regardless, this gets the basics implemented.