add release notes for historic releases
The notes about releases from v0.5.1 to v0.8.0 are based on release
notes which existed in earlier commits.

The notes about releases from v0.8.1 to v1.2.0 have been compiled now.
For v0.8.1, v0.8.2, v1.1.1, v1.1.2, v1.2.1, and v1.2.2 all commits have
been inspected.  For v1.0.0, v1.1.0, and v1.2.0 only the diffs comparing
with the previous release have been inspected.

There certainly are changes which should have been mentioned but are
not.  Feel free to complete the lists of changes and also to change the
wording to be more in line with the humanized style used in Git release
notes.