magit-file-relative-name: fix for W32 with Cygwin
On Windows, Emacs formats file names as e.g. c:/path/to/somewhere. This
causes a mismatch with Cygwin's git, which uses /path/to/somewhere.

Bridge these two by calling `expand-file-name' on the result of
`magit-get-top-dir' before calculating its length, so that it's in the
same format as the file's name. The file's name is in that format via
`file-truename', and also because it probably came from Emacs rather
than git.