magit-ediff-popup: new popup, replacing old bindings
Add the new popup `magit-ediff-popup' from which all ediff variants can
be invoked.  Bind `E' to that, and remove the old direct meta-<key>
bindings used for the individual commands.  These bindings were always
just kludges and I already intended to find a better solution when I
added them.  (As much as possible Magit should limit its keybindings
to letters).

The `magit-ediff-dwim' gets a new binding `e'.  Only `r' is now bound
to `magit-rebase-popup', but since we are running out of keys, having
two keys bound to the same command is no longer an option.
`magit-ediff-dwim' gets the better binding than `magit-ediff-popup'
because the former ... well dwims.  But not always which is why the
popup had to be added.