require at least version 1.9.4 of Git
This version was released a year ago for all platforms except
MS Windows.

The MS Window port which can be downloaded from
http://git-scm.com/download/windows never saw a release named 1.9.4,
instead something named 1.9.5 was finally released two weeks ago.
Previously the port was stuck on an even older release.  Git proper
itself never used the v1.9.5 tag - I don't know what's going on here.

So for MS Windows users the new minimal version now is the so-called
v1.9.5, for users on other platforms it is the actual v1.9.4.

I would have liked to go with v2.0.0, but cannot do so until an easily
installable port of that exists for MS Windows.