with-editor-post-cancel-hook: new hook
We will need to pass information from the pre-cancel to the post-cancel
hook, so add a variable to do that.

At the time the post-cancel hook is run, the buffer no longer exists, so
we have to preserve the local value using a let-binding.  Because we use
lexical scope the byte-compiler would complain that we don't use the
variable (which isn't true), so `with-editor-cancel-alist' has to be
declared globally.  Somehow the use of a let-binding causes the hook
functions to be run twice - seems like the let-bound value is considered
to be both the local as well as the global value.  So remove t from the
hook list.