require at least version 24.4 of Emacs
That release is the first to implement parts of the region overlay
handling in elisp.  The change in the next commit depend on the ability
to do just that; it allows decoupling hunk painting from section
highlighting.  And that in turn enables various significant performance
improvements.

So we have to do it - poor performance is most often voiced criticism.
This had to happen eventually.  There have been many small reasons why
dropping support for older Emacsen would have benefited the maintainer
and the majority of users.  Now there is a big one.  Note that one of
the reasons I released 1.4.0 was to have a recent release which still
works with older Emacsen.

Remove all the backward compatibility definitions which now are no
longer required.