avoid re-highlighting when moving within the same section
In `magit-section-update-highlight' check whether the current section is
the same as the one that was current before.  If that is so do nothing.
This makes moving inside diffs much cheaper.  We keep track of the
current section using the new variable `magit-section-highlighted-section'.

Because sibling selection overlays are highlighted during highlighting
and therefor not tracked as region overlays, their removal has to be
forced by setting `magit-section-highlighted-section' to nil.  Therefor
we might still do some unnecessary re-highlighting after deactivating
the region.