avoid un-/highlighting hidden hunks until expansion
When a hidden hunk section is selected it is no longer highlighted, and
when it is deselected it is no longer unhighlighted.  In many cases this
avoids unhighlighting and the rehighlighting again when moving back,
both of which would be useless because the affected part of the hunk is
not even visible.  This also affects un-/refining.

Instead hidden hunks that are not yet highlighted are highlighted when
they are expanded.

The heading of the hunk, which usually is visible, is un-/highlighted as
before, when ever the section is de-/selected.

When forgoing to unhighlight a hunk section
add it to `magit-section-highlighted-sections' and
remove it from `magit-section-unhighlight-sections'.

Hunks are still the only sections which are ever added to these
variables.  Which means that all other sections are always
un-/highlighted, even when that would not be necessary.

The old reason for why we don't avoid un- and rehighlighting other
sections is that doing so is believed to be cheap.  Implementing the
necessary housekeeping to avoid doing so might be more expensive even.

There is a new reason now: if an ancestor section of a hidden hunk
section whose body highlighting is not correct were marked were added to
the list of fully highlighted sections, then the delayed highlighting of
the hunk would fail.

All members of `magit-section-highlighted-sections' are known to be
fully and correctly highlighted.  Other sections might or might not
be fully and correctly highlighted, but we don't keep track of it.