avoid un-/refining invisible hunks until expansion
The new function `magit-diff-un-/refine-hunk' takes care of
everything now.

Also remove the variable `magit-diff-refine-hunk-related', which was
a temporary hack.  Like in the last release and before, it once more
is only possible to refine all hunks or the one and only current hunk.
Remove the option to refine all selected hunks.