avoid painting selected hunks twice after refresh
Previously these sections were first painted as unhighlighted and then
painted again as highlighted.

Never call `magit-diff-paint-hunk' directly in `magit-diff-wash-hunk'.
Instead rely on `magit-section-show' triggering the (second) painting.

That means that `magit-diff-paint-hunk' now has to be able to figure
out whether it is expected to highlight or unhighlight.  That could be
somewhat expensive, so caller which know which is the case continue to
pass on that information.  But `magit-section-show' doesn't know.

When `magit-diff-paint-hunk' is called without the HIGHLIGHT argument
then it used the new `magit-section-selected-p' to figure out whether
it should highlight or unhighlight.

`magit-section-selected-p' needs to check whether the section it
is called on is a member of the multi-section selection, which is
expensive if we were to do it several times.  So the functions that
do unhighlighting now pass along that information, just like the
functions that highlight already did.