rename magit-rebase-onto to magit-rebase-from
This breaks consistency with Git's terminology, which is a good thing
because that does not make sense here.

What differs between these two:

  git rebase UPSTREAM BRANCH
  git rebase --onto NEWBASE "UPSTREAM" BRANCH

isn't that the later rebases `--onto' something; all rebases do.  The
difference is that in the former case everything not in UPSTREAM is
rebased, while in the latter case only a subset of the "local" commits
is rebased.  So NEWBASE is actually the same thing as UPSTREAM, and
"UPSTREAM" should actually be named START-POINT or something like that.
And instead of an `--onto' argument there should be a `--from' argument
or instead of a branch it should be possible to specify a range:

  git rebase UPSTREAM --from START BRANCH
  git rebase UPSTREAM START..BRANCH