use more semantic colors for reflog faces
Use the color to indicate how dangerous the logged action is.

Creating a new commit is totally safe.  Amending to an existing commit
is more dangerous as it might not only add to but also remove from the
commit.  And resetting is very dangerous as it throws away one or more
commits.

Like creating commits, checking out a commit is also safe, but because
there are notable differences we use another "safe color".

Also lose the background colors for consistency with refname faces,
which also no longer have a background color.