use overlays to hide section bodies
Previously we used text properties but we never managed to fully teach
Emacs what we mean by "invisible".  We mean "cannot be seen with the
eyes", as opposed to "position that has the `invisible' text property
set, when taking into account text property stickiness and whatever else
is relevant".  We did force the appropriate stickiness and added a post
command kludge to work around that not having the desired effect.  And
maybe there was a bug in that.  But I do suspect there is a bug in Emacs
itself with regards to "taking into account text property stickiness"
when moving to the edges of the abyss - but I am not sure.

Anyway, by using an overlay we bypass all that.  When using an overlay
the default rules seem to lead to "invisible" := "cannot be seen by the
eyes".  For the most part this is implemented by copying the respective
code from `outline-flag-region', so it is code that has been around for
a long time.

There is one difference; when hiding a section body, we do *not* hide
the newline at the end of the heading but *do* hide the newline at the
end of the body - `outline-flag-region' does the opposite. That means
that hiding the body does not remove the background color of parts of
the heading - and that we cannot display an ellipsis at the "end of the
heading" (that wouldn't be visible).

This change will also allow isearching invisible section bodies, but
that isn't enabled yet - it still needs some work.

Re #1771, and many issues before that.