work around a cygwin path issue
hopefully.

In `magit-expand-git-file-name' turn paths such as
- /cygdrive/c/the/repo
- /c/the/repo
into paths like
- /c:/the/repo

If it turns out that there are places where an absolute path is
requested from git, then we need to use this function there too.
But I think this function is currently used when appropriate.

This issue only exists, or so I assume, when using Cygwin Git with a
non-cygwin Emacs.  It this were the case, then it would be appropriate
to only use this kludge when the `system-type' is `windows-nt'.  But I
am not sure, so for now it is also used when it is `cygwin'.  In the
latter case it could be that Git returns `/c/the/repo' while Emacs
expects `/c:/the/repo'.

Re #1318.