magit-diff-paint-hunk: respect nil magit-diff-highlight-hunk-body
I.e. make sure that painting is triggered despite the fact that this
setting invalidates some of the assumptions in the code that follows.
Do that by explicitly overriding the HIGHLIGHT argument to always be
nil when that option is nil.  With that in place the assumptions
apply again.

Fixes the second issue in #1842.