consolidate the functions that return the repository toplevel
Remove the variants `magit-get-top-dir' and `magit-toplevel-safe' in
favor of just `magit-toplevel'.  This function now combines all the
special cases previously handled by the removed functions.

`magit-toplevel' differs from `git rev-parse --show-toplevel' which it
uses internally, in that it also works from inside the .git directory
and in bare repositories.  The new STRICT argument can be used to
suppress these fallbacks.

Also add the new `magit-with-toplevel' macro which just let-binds
`default-directory' to the toplevel directory.