magit buffers no longer visit truename of toplevel
Previously we used `git rev-parse --show-toplevel' to determine the
toplevel directory of the current repository and used that as
`default-directory' of Magit buffers.  `--show-toplevel' always chases
symlinks, and so we ended up visiting the truename of that directory.

That's not nice and so we now use `--show-cdup' and construct the
absolute filename from that and `default-directory' using
`expand-file-name'.

Also remove several kludges where filenames had to be normalized
using `file-truename' and similar, before being compared to the
`default-directory' of Magit buffers.

However a user may explicitly set `find-file-visit-truename' to t.
To cover that case `magit-mode-setup' now respects that variable
when first creating a buffer.