update the mode-line element vc-mode more efficiently
For file-visiting buffers the VC package show the file's status in the
mode-line.  We used to call the *function* `vc-find-file-hook' after
reverting such a buffer to help VC keep it up-to-date.  However that
function is terribly inefficient and does other things we usually don't
need.

Instead use a new function `magit-refresh-vc-mode-line' which is like
`vc-mode-line' but simpler, more efficient, and less buggy.

Then instead of calling that function unconditionally add it to the
relevant hooks; allowing users to disable even this more efficient
implementation.

By not using `vc-find-file-hook' we also stop calling `vc-follow-link'.
But that should not matter much, because that function is still called
when the buffer is first created.