add magit-insert-headers, use it in magit-insert-status-headers
`magit-insert-headers' allows inserting a user-editable list of sections
while maintaining the ability to use the first section as the parent of
the other sections.  Furthermore if the first section is canceled then
the next section is used as parent.

This requires a new support function `magit-insert-remaining-headers'
and two new support variables `magit-insert-section-hook' and
`magit-insert-headers-hook'.

Also add two new section inserters intended to be used on
`magit-insert-status-headers-hook'.

Re #1879.