add kludges for ambiguous refs
`magit-ref-fullname' now returns nil when NAME is ambiguous.

And in `magit-branch-delete' ask the user what she is trying to delete
when we are not sure.  Ambiguous refs are user mistakes and we should
allow fixing these mistakes using Magit.  But anything that goes beyond
that would not be worth it because it would force users to deal with
full refs all the time in many places where we would rather use the
short branch names.

Fixes #1888.