magit-diff-update-hunk-refinement: add mouse kludge
When clicking somewhere using the mouse, then Emacs very briefly
considers the region to be active, even if one is careful not to
drag.  That's probably useful for something, but because we visualize
hunk-internal region in a way that can even show an empty region that
caused it to be briefly visible.

Add a kludge to prevent the flashing of the hunk-internal region.

Re #1891.