magit-expand-git-file-name: fix handling of remote paths
For remote paths `magit-expand-git-file-name' returned a path which
duplicated the remote part.  Now it always returns the complete path;
when it is a remote path, including the remote part (but only once).
The LOCALNAME argument is removed from this function as well as
`magit-git-path'.  The only caller of the latter which used to use
that argument, `magit-with-blob', now strips the remote part itself.