ediff: cleanup auxiliary buffers
Previously the intention was that `ediff-cleanup-mess' would take care
of this.  That function expects that the control buffer is current when
it runs, but our own hook function, which runs firsts, restores the
pre-ediff window configuration, making that buffer non-current.  So
`ediff-cleanup-mess' ended up doing nothing.

Reversing the order does not work because our function also expects that
the control buffer is current.  Also `ediff-cleanup-mess' does things to
the file-visiting buffers that we are not interested in.

So instead set the local value of `ediff-quit-hook' to a single
function, our own cleanup lambda.  This means that the global hook value
is ignored, including `ediff-cleanup-mess'.  `smerge-ediff' does this
too.  However that function directly calls `ediff-quit-hook' while we
call the trimmed down variant `magit-ediff-cleanup-auxiliary-buffers'.

Re #1934.