magit-log-cutoff-length: increase default value
The intention when setting this to allow value, is to make it faster.
However the problem isn't that Git is slow at outputting the log or
Magit is slow at parsing the output, the problem is that `git log
--graph' has a constant overhead which does not depend on how many
commits are being displayed but the total number of commits.

Therefore if this number is to low, it becomes more likely that a
user has to endure that delay several times until doubling the number
of displayed commits until finally all commits which she wants to be
listed are indeed listed.