improve regular expression for matching graph part of log
Matching the graph part of a log (generated with --graph)
takes a very long time. A post about how emacs deals with
regular expressions revealed that one should avoid nested
repition operators. Apparently the manual was updated to
explain this, the explanation is after the description of
the '*' special character.

Instead, I used a single repetition operator and moved the
space match inside the character selector. This works so
far, but I haven't tested this extensively...