magit-section-secondary-heading: new face
Add a new face `magit-section-secondary-heading', which is intended for
sub-sections which themselves have children.

The children of the "Bisect Log" section are such sections, their
heading looks like "git bisect ACTION HASH" and their bodies are hidden
by default.  It is non-obvious that these sections have a body so we
have to provide a hint.  Now their headings are bold, just like those
of `file' sections, which, as anyone should know, can be expanded.