magit-blame-mode: refuse to activate when run interactively
The commands `magit-blame' and `magit-blame-popup' would be used to
turn on this mode.  Previously turning on the mode directly also
worked, because the mode function then called `magit-blame', but
because that in turn calls the mode function, that ended up being
called twice.  And that resulted in the buffer becoming read-only.

Since `magit-blame-mode' was not intended to be called interactively
we now just raise an error (with usage instructions) if the user
attempts to do so.

Fixes #1948.