don't color log graph by default
Colorful graphs are nice, but they are just eye-candy.  At the same
time they lead to a delay because the control characters have to be
turned into faces using `ansi-color-apply-on-region'.

So stop hard-coding the `--color' argument.  Instead make it available
from the log popup and allow adding it to `magit-log-section-args'.