magit-diff-type: try harder in magit-diff-mode buffers
The only situation in which we cannot determine a proper type inside a
`magit-diff-mode' buffer and instead return `undefined' is when showing
the differences between the index and a commit that isn't `HEAD'.  There
is no `committed+index' type.  Previously this function never returned
`committed' when called in a `magit-diff-mode' buffer.