magit-diff-hunk-region-patch: do less
`magit-diff-hunk-region-patch' used to return the concatenation
of the file header and the adjusted hunk header, now it returns
just the latter.  So `magit-apply-region' now also has to call
`magit-diff-file-header' directly.

`magit-diff-hunk-region-patch' also used to perform some sanity
checks useful when applying a patch.  But since it is also used
by `magit-diff-hunk-region-header' which in turn is used to
adjust a hunk's heading when there is a hunk internal region,
this should not be done here.  Move the checks to
`magit-apply-region'.

Adjust `magit-diff-hunk-region-header' accordingly.