store complete diff headers in section structs
Previously we stored just the blobs in the BLOBS slot and later when the
file or parts of it were applied we constructed an (incomplete) header
in `magit-diff-file-header' using the BLOBS, FILE, and ORIG slots.  That
didn't work on new, deleted, and renamed files, because some information
was left out which is required in these cases.

Now just store the complete header in the new DIFF-HEADERS (which was
created by re-purposing and renaming BLOBS, which is no longer needed).