WIP do not indent revision summaries
This only serves to demonstrate what this could look like.  The existing
code is completely wacky and the changes only make it worse.  But this
gives a rough idea of what the result could like (in the revision
buffer, not the code!).  Where you see italic, pretend you see a color
slightly different from the foreground of `default' instead.  You need
to kill the revision buffer after loading this code, or it won't fully
take effect.