restore magit-branch-manager to ease users into the new binding
A lot of users where having problems finding the successor of the
Branch Manager (as well as the Wazzup buffer, but nobody seems to
miss that).

So restore the old key binding.  Even add a command named
`magit-branch-manager'.  But now it's just a wrapper around
`magit-show-refs-head' which instructs users to use the `y'
instead.