refs buffer: always mark the current branch, even without focus
When the commit counts and cherries are displayed relative to something
other than the current branch, then still mark it, using ".".  When the
focus is on the current branch, then continue marking it with "@" as
before.  When the focus is on something other than the current branch,
then mark it with "#", as before.