refs buffer: RET on a ref checks out
Starting with v2.1.0, "RET" no longer checked out the reference at
point, instead it was shown using `magit-show-commit'.  I have decided
to restore the old behavior, which unfortunately requires the below
hack because the references in this buffer use the same section types
`branch' and `tag' as in other buffers, and `magit-show-commit' is
bound in the respective section type keymaps.

So replace that command with `magit-visit-ref' in the those keymaps.
Except in `magit-refs-mode' buffers this command simply calls the
regular `magit-show-commit' command.  In refs buffers, on a reference,
it instead checks out that reference, like RET used to do.  With a
prefix argument it instead just focuses on that reference.

Re #2009 and #2062.