don't bind default-directory to pass topdir to magit-mode-setup
In `magit-status-internal' do not let-bind `default-directory' to "pass"
the top-level directory to `magit-mode-setup'.  Generally doing so would
be okay for this particular variable, while one should of course usually
use function arguments to pass values.

In this particular case binding `default-directory' confuses
`magit-mode-get-buffer' which is used by `magit-mode-setup'.  The result
would be that it were no longer possible to use `C-u C-u C-x g' from
within a repository to switch to another repository whose top-level
directory has the same basename as that of the current repository, when
`magit-status-buffer-name-format' contains `%b', i.e. when we only show
the basename in the mode-line.  It would think that we are already in
that buffer (because the dynamic binding of `default-directory' appears
to claim that we are).

So instead we should pass the path to the repository to be visited by
passing it to `magit-mode-setup' using a new function argument.  But
that's not possible because it would break existing code. We have to
use a new variable `magit-mode-setup--topdir' to pass the value to
`magit-mode-setup'.

Re #2054 and #2060.