magit-git-command: fix displayed process buffer
Previously, when called from a subdirectory, magit-git-command showed
one process buffer, while actually creating and dumping output to
another.  This was because magit-git-command displayed a buffer with the
name with the root directory and the default-directory set to the root
directory, but, when magit-run-git-async was called from the
subdirectory, it tried to grab a buffer with the same name but a
different default-directory.  This mismatch resulted in a second process
buffer being created (and not displayed).

* Change magit-git-command to create a process buffer with the
  subdirectory in the name and the correct value for default-directory.

* Add a default-directory argument to magit-process-setup (matching
  magit-process-finish) so that it too includes the subdirectory in the
  buffer name.