make magit-version even more defensive
If we conclude that one of the three supported installation methods was
used, then try to determine the version using the appropriate approach,
like we always did.  But if that fails, then give the other approaches a
change too.  Also collect debugging information and show it if we cannot
determine the version.