define hooks like we are supposed too
Users are allowed to use `add-hook' et al. to add hook functions to hook
variables which have not yet been declared.  That means that the default
value of hook variables should not be set inside `defvar' and defcustom'
forms, because those definitions are only evaluated if the variable has
no value yet (in case of `defcustom' it would be used to get the default
value though).  When the user used `add-hook' earlier, then it obviously
has a value.  So we are supposed to set the default value outside of the
declaration form, using `add-hook'.  But that has a few downsides too.

If we set the default value outside of `defcustom', then Custom would
think that the default value is `nil' (and would be right about that)
and tell the user about it, which could confuse her a lot.

So we set the default value inside `defcustom' *and* outside using
`add-hook'. Sight.  That's still not good enough though.  Re-evaluating
the file or buffer containing these `add-hook' forms would override all
user customization.

So we place the calls to `add-hook' before that to `defcustom' and
additionally wrap the calls to `add-hook' to make sure they are only
ever run once.  Sight.

(unless (find-lisp-object-file-name 'HOOK-VARIABLE 'defvar) ...)
  (add-hook 'HOOK-VARIABLE ...) ...)
(defcustom HOOK-VARIABLE '(...)
  :options '(...)
  ...)

Double sight.

We *don't* do the above for section hooks, which users are supposed to
modify using `magit-add-section-hook'.  Instead we teach that function
to raise an error when the user attempts to add a hook function to a
hook variable which has not been defined yet.

Re #2004.