magit-mode-get-buffer: fix sub-directory case
Previously a caller could set TOPDIR argument, so that this function
could avoid calling `magit-toplevel' if the caller already knew that
value.  But the assumption that we could do that was wrong: we cannot
get around calling that function because in some cases we actually want
to run git in a directory which is a sub-directory of a repository, not
a repository top-directory (or outside any repository at all).

Rename TOPDIR to DIRECTORY.  It's only purpose now is to override
`default-directory'.  Always call `magit-toplevel' and find the buffer
for that repository, instead of defaulting to use a non-nil DIRECTORY.

Re #2100.