replace TOPDIR argument with PWD in process buffer getters
Rename the TOPDIR arguments of `magit-mode-get-process',
`magit-mode-get-process-create', and `magit-process-buffer' to PWD.

The original intention was that callers could pass TOPDIR, in
order to avoid calling it again inside `magit-mode-get-buffer' when
its value is already known.  However in practice TOPDIR is in some
cases not actually the top-level directory but some sub-directory of
the repository.

In `magit-mode-get-process' always call `magit-toplevel' and find or
create a buffer with a matching `default-directory'.  Take non-nil PWD
into account by let-binding `default-directory' to it while determining
the top-level.