magit-process-insert-section: add mandatory PWD argument
The `default-directory' of a process buffer is the top-level directory
of the corresponding repository, but it should also be possible to run
git in a sub-directory without having to let-bind `default-directory'.

When `magit-process-insert-section' is called then the process buffer
always is the current buffer.  This made it impossible to pass the
working directory to this function except by let-binding
`default-directory' inside the process buffer, which had some rather
nasty side-effects.  The new PWD argument makes it possible to do so
without causing any problems.