magit-git-command: don't call git inside the process buffer
`magit-git-command' may run git in any directory, not just the top-level
directory.  Previously we made the process buffer current and then
let-bound `default-directory' to the subdirectory, but that confused
`magit-process-buffer' (called by `magit-run-git-async').  It was able
to find the correct process buffer by name, but then it concluded that
the `default-directory' was wrong because it did not match
`magit-toplevel'. The reason for that is that the let-binding was still
in effect.

So instead call `magit-run-git-async-args' first, with the
`default-directory' of whatever buffer is current in effect.  And only
then call `magit-mode-display-buffer'.  Let-binding `default-directory'
inside a process buffer is now considered a bug (but this was the only
place where we did it).