magit-diff-arguments: always include the files in return value
The function `magit-diff-arguments' now always returns a list with two
elements, the list of regular arguments and the list of file arguments.
Commands that use that function, and didn't do so already, are adjusted
to accept that additional argument and pass it along.  The functions
that they pass it along to are adjusted accordingly, which in some
cases does not have an effect yet.